refactor: 파일명 및 함수명 및 className에 카멜케이스를 적용해서 수정
