+++ src/components/FlotingButtons/SignInButton/SignInButton.module.scss
@@ -3,7 +3,7 @@
     color: var(--color-text);
   }
 
-  .signinButton {
+  .signInButton {
     color: var(--color-text);
   }
 
@@ -27,7 +27,7 @@
           background-color: var(--color-bg1);
         }
 
-        .signoutButton {
+        .signOutButton {
           color: var(--color-text);
         }
       }
