+++ src/components/FlotingButtons/SignInButton/index.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import useModalStore from '@/store/zustand/useModalStore';
 import useStore from '@/store/zustand/useStore';
 import useAuthStore from '@/store/zustand/useAuthStore';
 import Typography from '@/components/Typography';
-import { signoutApi } from '@/service/sign';
-import styles from './loginButtons.module.scss';
+import { signOutApi } from '@/service/sign';
+import styles from './SignInButton.module.scss';
 
-export default function LoginButton() {
+export default function SignInButton() {
   const [isDropdownVisible, setDropdownVisible] = useState<boolean>(false);
 
   const setModal = useModalStore((state) => state.setModal);
@@ -28,7 +28,7 @@ export default function LoginButton() {
     try {
       if (userInfo && setUserInfoInit) {
         const param = { userId: userInfo.userid };
-        await signoutApi(param);
+        await signOutApi(param);
         setUserInfoInit();
         setDropdownVisible(false);
       }
@@ -58,7 +58,7 @@ export default function LoginButton() {
           <Typography variant="body2">{userInfo.email}</Typography>
         </button>
       ) : (
-        <button className={styles.signinButton} type="button" onClick={handleModal}>
+        <button className={styles.signInButton} type="button" onClick={handleModal}>
           <Typography variant="body2"> 로그인</Typography>
         </button>
       )}
@@ -67,7 +67,7 @@ export default function LoginButton() {
         <div className={styles.dropdown} ref={dropdownRef}>
           <ul className={styles.ul}>
             <li className={styles.li}>
-              <button type="button" className={styles.signoutButton} onClick={handleSignOutClick}>
+              <button type="button" className={styles.signOutButton} onClick={handleSignOutClick}>
                 <Typography variant="body2">로그아웃</Typography>
               </button>
             </li>
