+++ src/components/FlotingButtons/index.tsx
@@ -1,5 +1,5 @@
 import styles from './floatingButtons.module.scss';
-import LoginButton from './LoginButton';
+import LoginButton from './SignInButton';
 import ToggleThemeButton from './ToggleThemeButton';
 
 export default function FloatingButtons() {
